requirements.txt
XlsxWriter
dash
plotly
lightgbm
numpy
pandas
matplotlib
google-cloud-storage
gunicorn
openpyxl
scikit-learn